nl36_extractor/requirements.txt
pdfplumber>=0.9.0
openpyxl>=3.1.0
pyyaml>=6.0
